requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
backtrader==1.9.78.123
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
executing==2.1.0
fonttools==4.55.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipython==8.31.0
jedi==0.19.2
jiter==0.8.2
kiwisolver==1.4.8
matplotlib==3.10.0
matplotlib-inline==0.1.7
numpy==2.2.1
openai==1.59.6
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scipy==1.15.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
